page location refactor
